Tighten argument about A() injectivity vs K(), X()
It doesn't follow that *both* K and X are injective, from the fact that A is injective, in A(mu) = (X(mu), K(mu)), it only follows that one of them are. But Schnorr is essentially symmetric in x, k so you can just prove it for one of the two and the other is trivial.