Correct argument in Theorem 2
Previously there was an erroneous framing of h_e(x) = g(x) + ex, which at best was highly confusing but more realistically was just wrong: e is a function of x, itself. The argument for injectivity is now more straightfowardly logical and we only refer to h(x) not h_e(x).